add validation for the rest of the functions in CelestialObjectService

+++ src/main/java/com/matus/spacecatalog_backend/service/CelestialObjectService.java
@@ -3,6 +3,7 @@ package com.matus.spacecatalog_backend.service;
 
 import com.matus.spacecatalog_backend.entity.CelestialObject;
 import com.matus.spacecatalog_backend.repository.CelestialObjectRepository;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -20,22 +21,50 @@ public class CelestialObjectService {
     @Autowired
     private CelestialObjectRepository repository;
 
+    @Transactional
     public CelestialObject create( CelestialObject object) {
+
+        if(object.getObjectName() == null || object.getObjectName().isBlank()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Object name cant be null or empty");
+        }
+
+        repository.findByObjectName(object.getObjectName()).ifPresent(existing -> {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "Celestial name already exists");
+        });
+
+        if(object.getObjectDesignation() != null || !object.getObjectDesignation().isBlank()){
+            repository.findByObjectDesignation(object.getObjectDesignation()).ifPresent(existing -> {
+                throw new ResponseStatusException(HttpStatus.CONFLICT, "Celestial designation already exists");
+            });
+        }
+
        return repository.save(object);
     }
 
+    @Transactional(readOnly = true)
     public Page<CelestialObject> findAll(Pageable pageable) {
         return repository.findAll(pageable);
     }
 
-    public Optional<CelestialObject> findById(Long id) {
-        return repository.findById(id);
+    @Transactional(readOnly = true)
+    public CelestialObject findById(Long id) {
+
+        CelestialObject existingObject = repository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Celestial object not found by ID"));
+
+        return existingObject;
     }
 
+    @Transactional
     public void delete(Long id) {
+        if(!repository.existsById(id)){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Not found for delete");
+        }
+
         repository.deleteById(id);
     }
 
+    @Transactional
     public CelestialObject updateObject (Long id, CelestialObject updatedData){
 
         CelestialObject existing = repository.findById(id)
@@ -73,8 +102,14 @@ public class CelestialObjectService {
 
     }
 
+    @Transactional(readOnly = true)
     public Page<CelestialObject> findByType(String objectType, Pageable pageable){
+
+        if(objectType == null || objectType.isBlank()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Object type cant be null or empty");
+        }
+
         return repository.findAllObjectType(objectType, pageable);
     }
-    
+
 }
